+++ pages/manageFormula.php
@@ -126,6 +126,31 @@ if($_GET['formula'] && $_GET['do']){
 	}
 	return;
 	
+//ADD NEW FORMULA
+}elseif($_POST['action'] == 'addFormula'){
+	if(empty($_POST['name'])){
+		echo '<div class="alert alert-danger alert-dismissible"><strong>Formula name is required.</strong></div>';
+		return;
+	}
+	$name = mysqli_real_escape_string($conn, $_POST['name']);
+	$notes = mysqli_real_escape_string($conn, $_POST['notes']);
+	$profile = mysqli_real_escape_string($conn, $_POST['profile']);
+	$fid = base64_encode($name);
+	
+	if(mysqli_num_rows(mysqli_query($conn, "SELECT fid FROM formulasMetaData WHERE fid = '$fid'"))){
+		echo '<div class="alert alert-danger alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><strong>Error: </strong>'.$name.' already exists! Click <a href="?do=Formula&name='.name.'">here</a> to view/edit!
+			</div>';
+		}else{
+			$q = mysqli_query($conn, "INSERT INTO formulasMetaData (fid, name, notes, profile, image) VALUES ('$fid', '$name', '$notes', '$profile', '$def_app_img')");
+				if($q){
+					echo '<div class="alert alert-success alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><strong><a href="?do=Formula&name='.$name.'">'.$name.'</a></strong> added!</div>';
+				}else{
+					echo '<div class="alert alert-danger alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><strong>Something went wrong...</strong></div>';
+				}
+		}
+
+	return;
+	
 	
 //DELETE FORMULA
 }elseif($_GET['action'] == 'delete' && $_GET['fid']){